third_party/WebKit/LayoutTests/fast/dom/html-options-collection-lifetime.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<body>
<script>
description('Tests that HTMLOptionsCollection does not keep its owner node alive.');

var select = document.createElement("select");
var option = document.createElement("option");
select.appendChild(option);
var options = select.options;
var observeSelect = internals.observeGC(select);
var observeOption = internals.observeGC(option);
var observeOptions = internals.observeGC(options);

select = null;
option = null;
gc();

shouldBeTrue('observeSelect.wasCollected');
shouldBeTrue('observeOption.wasCollected');
shouldBeFalse('observeOptions.wasCollected');
</script>
</body>
</html>
