third_party/WebKit/LayoutTests/fast/dom/html-options-collection-lifetime-expected.txt

Tests that HTMLOptionsCollection does not keep its owner node alive.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS observeSelect.wasCollected is true
PASS observeOption.wasCollected is true
PASS observeOptions.wasCollected is false
PASS successfullyParsed is true

TEST COMPLETE
